refactor:DogAnimationの画像 表示用のコンポーネント作成,components/animations内の構成変更、DogAnimationのリファクタリング完了
