+++ components/animations/puppyCat/PuppyCatAnimation.tsx
@@ -2,18 +2,18 @@ import React, { useEffect, useRef, useState } from 'react';
 import gsap from 'gsap';
 import Image from 'next/image';
 
-import legImageBackRight from '../../public/三毛猫奥後足.png';
-import legImageFrontRight from '../../public/三毛猫奥前足.png';
-import legImageFrontLeft from '../../public/三毛猫手前前足.png';
-import legImageBackLeft from '../../public/三毛猫手間後ろ足.png';
-import tailImage from '../../public/三毛猫尻尾.png';
-import faceImage from '../../public/三毛猫顔.png';
-import bodyImage from '../../public/三毛猫胴体.png';
-import earImage from '../../public/三毛猫耳.png';
-import beardImageRight from '../../public/三毛猫右ひげ.png';
-import beardImageLeft from '../../public/三毛猫左ひげ.png';
+import legImageBackRight from '../../../public/三毛猫奥後足.png';
+import legImageFrontRight from '../../../public/三毛猫奥前足.png';
+import legImageFrontLeft from '../../../public/三毛猫手前前足.png';
+import legImageBackLeft from '../../../public/三毛猫手間後ろ足.png';
+import tailImage from '../../../public/三毛猫尻尾.png';
+import faceImage from '../../../public/三毛猫顔.png';
+import bodyImage from '../../../public/三毛猫胴体.png';
+import earImage from '../../../public/三毛猫耳.png';
+import beardImageRight from '../../../public/三毛猫右ひげ.png';
+import beardImageLeft from '../../../public/三毛猫左ひげ.png';
 
-import { usePuppyCatWalkingAnimation } from '../../hooks/components/animations/puppyCat/usePuppyCatWalkingAnimation'
+import { usePuppyCatWalkingAnimation } from '../../../hooks/components/animations/puppyCat/usePuppyCatWalkingAnimation'
 
 const PuppyCatAnimation= () => {
   const legBackLeftRef = useRef<HTMLImageElement | null>(null);
